+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Changed
+
+- [CL-1776] Remove feedback icon in Bottom Right
+
 ## 2022-11-08
 
 ### Added